+++ mock/app.go
@@ -6,16 +6,18 @@ import (
 	"path/filepath"
 
 	abci "github.com/tendermint/abci/types"
+	crypto "github.com/tendermint/go-crypto"
+	tmtypes "github.com/tendermint/tendermint/types"
 	dbm "github.com/tendermint/tmlibs/db"
 	"github.com/tendermint/tmlibs/log"
 
 	bam "github.com/cosmos/cosmos-sdk/baseapp"
 	sdk "github.com/cosmos/cosmos-sdk/types"
+	"github.com/cosmos/cosmos-sdk/wire"
 )
 
-// NewApp creates a simple mock kvstore app for testing.
-// It should work similar to a real app.
-// Make sure rootDir is empty before running the test,
+// NewApp creates a simple mock kvstore app for testing. It should work
+// similar to a real app. Make sure rootDir is empty before running the test,
 // in order to guarantee consistent results
 func NewApp(rootDir string, logger log.Logger) (abci.Application, error) {
 	db, err := dbm.NewGoLevelDB("mock", filepath.Join(rootDir, "data"))
@@ -103,11 +105,10 @@ func InitChainer(key sdk.StoreKey) func(sdk.Context, abci.RequestInitChain) abci
 	}
 }
 
-// GenInitOptions can be passed into InitCmd,
-// returns a static string of a few key-values that can be parsed
-// by InitChainer
-func GenInitOptions(args []string, addr sdk.Address, coinDenom string) (json.RawMessage, error) {
-	opts := []byte(`{
+// AppGenState can be passed into InitCmd, returns a static string of a few
+// key-values that can be parsed by InitChainer
+func AppGenState(_ *wire.Codec, _ []json.RawMessage) (appState json.RawMessage, err error) {
+	appState = json.RawMessage(`{
   "values": [
     {
         "key": "hello",
@@ -119,5 +120,16 @@ func GenInitOptions(args []string, addr sdk.Address, coinDenom string) (json.Raw
     }
   ]
 }`)
-	return opts, nil
+	return
+}
+
+// Return a validator, not much else
+func AppGenTx(_ *wire.Codec, pk crypto.PubKey) (
+	appGenTx, cliPrint json.RawMessage, validator tmtypes.GenesisValidator, err error) {
+
+	validator = tmtypes.GenesisValidator{
+		PubKey: pk,
+		Power:  10,
+	}
+	return
 }