+++ server/init_test.go
@@ -8,17 +8,36 @@ import (
 	"github.com/tendermint/tmlibs/log"
 
 	"github.com/cosmos/cosmos-sdk/mock"
+	"github.com/cosmos/cosmos-sdk/wire"
 	tcmd "github.com/tendermint/tendermint/cmd/tendermint/commands"
 )
 
-func TestInit(t *testing.T) {
+// TODO update
+func TestInitCmd(t *testing.T) {
 	defer setupViper(t)()
 
 	logger := log.NewNopLogger()
 	cfg, err := tcmd.ParseConfig()
 	require.Nil(t, err)
 	ctx := NewContext(cfg, logger)
-	cmd := InitCmd(mock.GenInitOptions, ctx)
+	cdc := wire.NewCodec()
+	appInit := AppInit{
+		AppGenState: mock.AppGenState,
+		AppGenTx:    mock.AppGenTx,
+	}
+	cmd := InitCmd(ctx, cdc, appInit)
 	err = cmd.RunE(nil, nil)
 	require.NoError(t, err)
+}
+
+func TestGenTxCmd(t *testing.T) {
+	// TODO
+}
+
+func TestSimpleAppGenTx(t *testing.T) {
+	// TODO
+}
+
+func TestSimpleAppGenState(t *testing.T) {
+	// TODO
 }