+++ server/start_test.go
@@ -10,6 +10,7 @@ import (
 	"github.com/stretchr/testify/require"
 
 	"github.com/cosmos/cosmos-sdk/mock"
+	"github.com/cosmos/cosmos-sdk/wire"
 	"github.com/tendermint/abci/server"
 	tcmd "github.com/tendermint/tendermint/cmd/tendermint/commands"
 	"github.com/tendermint/tmlibs/log"
@@ -25,7 +26,12 @@ func TestStartStandAlone(t *testing.T) {
 	cfg, err := tcmd.ParseConfig()
 	require.Nil(t, err)
 	ctx := NewContext(cfg, logger)
-	initCmd := InitCmd(mock.GenInitOptions, ctx)
+	cdc := wire.NewCodec()
+	appInit := AppInit{
+		AppGenState: mock.AppGenState,
+		AppGenTx:    mock.AppGenTx,
+	}
+	initCmd := InitCmd(ctx, cdc, appInit)
 	err = initCmd.RunE(nil, nil)
 	require.NoError(t, err)
 
@@ -51,13 +57,18 @@ func TestStartWithTendermint(t *testing.T) {
 	cfg, err := tcmd.ParseConfig()
 	require.Nil(t, err)
 	ctx := NewContext(cfg, logger)
-	initCmd := InitCmd(mock.GenInitOptions, ctx)
+	cdc := wire.NewCodec()
+	appInit := AppInit{
+		AppGenState: mock.AppGenState,
+		AppGenTx:    mock.AppGenTx,
+	}
+	initCmd := InitCmd(ctx, cdc, appInit)
 	err = initCmd.RunE(nil, nil)
 	require.NoError(t, err)
 
 	// set up app and start up
 	viper.Set(flagWithTendermint, true)
-	startCmd := StartCmd(mock.NewApp, ctx)
+	startCmd := StartCmd(ctx, mock.NewApp)
 	startCmd.Flags().Set(flagAddress, FreeTCPAddr(t)) // set to a new free address
 	timeout := time.Duration(5) * time.Second