+++ server/util.go
@@ -1,12 +1,16 @@
 package server
 
 import (
+	"encoding/json"
+	"net"
 	"os"
 
+	"github.com/pkg/errors"
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
 
 	"github.com/cosmos/cosmos-sdk/version"
+	"github.com/cosmos/cosmos-sdk/wire"
 	tcmd "github.com/tendermint/tendermint/cmd/tendermint/commands"
 	cfg "github.com/tendermint/tendermint/config"
 	"github.com/tendermint/tmlibs/cli"
@@ -31,7 +35,7 @@ func NewContext(config *cfg.Config, logger log.Logger) *Context {
 	return &Context{config, logger}
 }
 
-//--------------------------------------------------------------------
+//___________________________________________________________________________________
 
 // PersistentPreRunEFn returns a PersistentPreRunE function for cobra
 // that initailizes the passed in context with a properly configured
@@ -62,18 +66,71 @@ func PersistentPreRunEFn(context *Context) func(*cobra.Command, []string) error 
 
 // add server commands
 func AddCommands(
-	rootCmd *cobra.Command,
-	appState GenAppState, appCreator AppCreator,
-	context *Context) {
+	ctx *Context, cdc *wire.Codec,
+	rootCmd *cobra.Command, appInit AppInit,
+	appCreator AppCreator) {
 
-	rootCmd.PersistentFlags().String("log_level", context.Config.LogLevel, "Log level")
+	rootCmd.PersistentFlags().String("log_level", ctx.Config.LogLevel, "Log level")
 
 	rootCmd.AddCommand(
-		InitCmd(appState, context),
-		StartCmd(appCreator, context),
-		UnsafeResetAllCmd(context),
-		ShowNodeIDCmd(context),
-		ShowValidatorCmd(context),
+		InitCmd(ctx, cdc, appInit),
+		StartCmd(ctx, appCreator),
+		UnsafeResetAllCmd(ctx),
+		ShowNodeIDCmd(ctx),
+		ShowValidatorCmd(ctx),
 		version.VersionCmd,
 	)
+}
+
+//___________________________________________________________________________________
+
+// append a new json field to existing json message
+func AppendJSON(cdc *wire.Codec, baseJSON []byte, key string, value json.RawMessage) (appended []byte, err error) {
+	var jsonMap map[string]json.RawMessage
+	err = cdc.UnmarshalJSON(baseJSON, &jsonMap)
+	if err != nil {
+		return nil, err
+	}
+	jsonMap[key] = value
+	bz, err := wire.MarshalJSONIndent(cdc, jsonMap)
+	return json.RawMessage(bz), err
+}
+
+// https://stackoverflow.com/questions/23558425/how-do-i-get-the-local-ip-address-in-go
+// TODO there must be a better way to get external IP
+func externalIP() (string, error) {
+	ifaces, err := net.Interfaces()
+	if err != nil {
+		return "", err
+	}
+	for _, iface := range ifaces {
+		if iface.Flags&net.FlagUp == 0 {
+			continue // interface down
+		}
+		if iface.Flags&net.FlagLoopback != 0 {
+			continue // loopback interface
+		}
+		addrs, err := iface.Addrs()
+		if err != nil {
+			return "", err
+		}
+		for _, addr := range addrs {
+			var ip net.IP
+			switch v := addr.(type) {
+			case *net.IPNet:
+				ip = v.IP
+			case *net.IPAddr:
+				ip = v.IP
+			}
+			if ip == nil || ip.IsLoopback() {
+				continue
+			}
+			ip = ip.To4()
+			if ip == nil {
+				continue // not an ipv4 address
+			}
+			return ip.String(), nil
+		}
+	}
+	return "", errors.New("are you connected to the network?")
 }